# labs/lab05-api-client-and-db-intg/starter/src/server.js
import express from "express";
import cors from "cors";
import pg from "pg";

const { Pool } = pg;

const PORT = process.env.PORT || 3000;

const pool = new Pool({
  host: process.env.PGHOST ?? "127.0.0.1",
  port: Number(process.env.PGPORT ?? 5433),
  database: process.env.PGDATABASE ?? "lab05",
  user: process.env.PGUSER ?? "postgres",
  password: process.env.PGPASSWORD ?? "postgres"
});

export function createApp() {
  const app = express();

  app.use(express.json());

  app.use(cors({
    /* 
    origin: [
    //   "http://localhost:5173",
    //   "http://127.0.0.1:5173",
    //   "http://192.168.1.*:5173",
    //   "http://172.18.0.1:5173",
    //   "http://localhost:3000",
    //   "http://127.0.0.1:3000"
     ]

     */

    origin: true // i know this is bad practice but i am using this with my swagger doc since my swagger plugin within vscode hates cors

  }));

  app.get("/health", async (req, res) => {
    try {
      await pool.query("SELECT 1");
      res.json({ status: "ok" });
    } catch (error) {
      console.error("Health check failed:", error);
      res.status(500).json({
        status: "error",
        message: "Database connection failed."
      });
    }
  });

  // Starter route: return every item from the database.
  app.get("/api/items", async (req, res) => {
    try {
      const result = await pool.query(`
        SELECT id, name, quantity, restock_quantity
        FROM items
        ORDER BY id ASC
      `);

      res.json({ items: result.rows });
    } catch (error) {
      console.error("Failed to load items:", error);
      res.status(500).json({
        error: "Internal Server Error",
        message: "Failed to load items."
      });
    }
  });

  // Starter route: create one item so the client can demonstrate a write.
  app.post("/api/items", async (req, res) => {
    const name = req.body?.name?.trim();
    const quantity = Number(req.body?.quantity);
    const restock_quantity = req.body?.restock_quantity

    if (!name || !Number.isInteger(quantity) || quantity < 0) {
      return res.status(400).json({
        error: "Bad Request",
        message: "A name and non-negative integer quantity are required."
      });
    }

    try {
      const result = await pool.query(
        `
          INSERT INTO items (name, quantity, restock_quantity)
          VALUES ($1, $2, $3)
          RETURNING id, name, quantity, restock_quantity
        `,
        [name, quantity, restock_quantity]
      );

      res.status(201).json({ item: result.rows[0] });
    } catch (error) {
      console.error("Failed to add item:", error);
      res.status(500).json({
        error: "Internal Server Error",
        message: "Failed to add item."
      });
    }
  });

  // DONE: Return one item by ID.
  app.get("/api/items/:id", async (req, res) => {
    const id = req.params.id;
    try {
      const result = await pool.query(`
        SELECT id, name, quantity, restock_quantity
        FROM items
        WHERE ID = $1
      `,
      [id]);
      if (result.rows.length === 0){
        res.status(404).json({ error: "Resource requested not found" });
      } else{
        res.json({ items: result.rows });
      }
    } catch (error) {
      console.error("Failed to load items:", error);
      res.status(500).json({
        error: "Internal Server Error",
        message: "Failed to load items."
      });
    }
  });

  app.get("/api/items/:id/restock", async (req, res) => {
    const id = req.params.id;
    try {
      const temp_result = await pool.query(`
        SELECT quantity, restock_quantity
        FROM items
        WHERE ID = $1
      `,
      [id]);
      if (temp_result.rows.length === 0){
        res.status(404).json({ error: "Resource requested not found" });
      } else{
        const new_quantity = temp_result.rows[0].restock_quantity + temp_result.rows[0].quantity;
        const middle_result =  await pool.query(
          `
          UPDATE items
          SET quantity = $1
          WHERE id = $2
          `,
          [new_quantity, id]
        );
        const result = await pool.query(`
        SELECT id, name, quantity, restock_quantity
        FROM items
        WHERE ID = $1
        `,
        [id])
        res.json({ items: result.rows });
      }
    } catch (error) {
      console.error("Failed to load items:", error);
      res.status(500).json({
        error: "Internal Server Error",
        message: "Failed to load items."
      });
    }
  });

  // DONE: Replace one item by ID.
  app.put("/api/items/:id", async (req, res) => {
    const name = req.body?.name?.trim();
    const quantity = Number(req.body?.quantity);
    const restock_quantity = Number(req.body?.restock_quantity);
    const id = req.params.id;

    if (!name || !Number.isInteger(quantity) || quantity < 0) {
      return res.status(400).json({
        error: "Bad Request",
        message: "A name and non-negative integer quantity are required."
      });
    }

    try {
      const result = await pool.query(`
        SELECT id, name, quantity, restock_quantity
        FROM items
        WHERE ID = $1
      `,
      [id]);
      if (result.rows.length === 0){
        res.status(404).json({ error: "Resource requested not found" });
      } else{
          try {
            const result =  await pool.query(
              `
              UPDATE items
              SET name = $1, quantity = $2, restock_quantity = $3
              WHERE id = $4
              `,
              [name, quantity, restock_quantity,  id]
            );

            res.status(201).json({ item: result.rows[0] });
          } catch (error) {
            console.error("Failed to add item:", error);
            res.status(500).json({
              error: "Internal Server Error",
              message: "Failed to add item."
            });
          }
            }
    } catch (error) {
      console.error("Failed to load items:", error);
      res.status(500).json({
        error: "Internal Server Error",
        message: "Failed to load items."
      });
    }


  });


  // DONE: Partially update one item by ID.
  app.patch("/api/items/:id", async (req, res) => {
    const id = req.params.id;
    try {
      const result = await pool.query(`
        SELECT id, name, quantity, restock_quantity
        FROM items
        WHERE ID = $1
      `,
      [id]);
      if (result.rows.length === 0){
        res.status(404).json({ error: "Resource requested not found" });
      } else{
              if ((req.body.hasOwnProperty('quantity') || req.body.hasOwnProperty('name')) || req.body.hasOwnProperty('restock_quantity') && Object.keys(req.body).length <= 3){
        if (req.body.hasOwnProperty('quantity')){
          const quantity = req.body.quantity;
          const result =  await pool.query(
            `
            UPDATE items
            SET quantity = $1
            WHERE id = $2
            `,
            [quantity, id]
          );
        }
        if (req.body.hasOwnProperty('name')){
          const name = req.body.name;
          const result =  await pool.query(
            `
            UPDATE items
            SET name = $1
            WHERE id = $2
            `,
            [name, id]
          );
        }
        if (req.body.hasOwnProperty('restock_quantity')){
          const restock_quantity = req.body?.restock_quantity
          const result =  await pool.query(
            `
            UPDATE items
            SET restock_quantity = $1
            WHERE id = $2
            `,
            [restock_quantity, id]
          );
        }

        try {
          const result = await pool.query(`
            SELECT id, name, quantity, restock_quantity
            FROM items
            WHERE ID = $1
          `,
          [id]);

          res.json({ items: result.rows });
        } catch (error) {
          console.error("Failed to load items:", error);
          res.status(500).json({
            error: "Internal Server Error",
            message: "Failed to load items."
          });
        }
      } else {
        res.status(400).json({ error: "Malformed json, please try again with only name and quantity" });
      } 
      }
    } catch (error) {
      console.error("Failed to load items:", error);
      res.status(500).json({
        error: "Internal Server Error",
        message: "Failed to load items."
      });
    }
  });

  // DONE: Delete one item by ID.
  app.delete("/api/items/:id", async (req, res) => {
    const id = req.params.id;

    try {
      const result = await pool.query(`
        SELECT id, name, quantity, restock_quantity
        FROM items
        WHERE ID = $1
      `,
      [id]);
      if (result.rows.length === 0){
        res.status(404).json({ error: "Resource requested not found" });
      } else{
          try {
            const result = await pool.query(`
              DELETE FROM items
              WHERE ID = $1
            `,
            [id]);
            res.status(204).json({ result: "Item successfully deleted" });
          } catch (error) {
            console.error("Failed to load items:", error);
            res.status(500).json({
              error: "Internal Server Error",
              message: "Failed to load items."
            });
          }
      }
    } catch (error) {
      console.error("Failed to load items:", error);
      res.status(500).json({
        error: "Internal Server Error",
        message: "Failed to load items."
      });
    }


  });

  app.use((req, res) => {
    res.status(404).json({ error: "Not found" });
  });

  return app;
}

export async function initializeDatabase() {
  await pool.query(`
    CREATE TABLE IF NOT EXISTS items (
      id SERIAL PRIMARY KEY,
      name TEXT NOT NULL,
      quantity INTEGER NOT NULL CHECK (quantity >= 0)
    )
  `);

  const { rows } = await pool.query("SELECT COUNT(*)::int AS count FROM items");

  if (rows[0].count === 0) {
    await pool.query(
      `
        INSERT INTO items (name, quantity)
        VALUES ($1, $2), ($3, $4), ($5, $6)
      `,
      ["Keyboard", 10, "Mouse", 5, "Monitor", 3]
    );
  }
}

const isMainModule = process.argv[1] === new URL(import.meta.url).pathname;

if (isMainModule) {
  const app = createApp();

  initializeDatabase()
    .then(() => {
      app.listen(PORT, () => {
        console.log(`Lab 5 API listening on http://localhost:${PORT}`);
      });
    })
    .catch((error) => {
      console.error("Server startup failed:", error);
      process.exit(1);
    });
}